test(widget-label): nest spec + extract helpers; drop repo-external doc ref
Move widgetLabelPersistence.spec.ts into browser_tests/tests/widgets/ and
extract setup constants + graph-state helpers into
fixtures/utils/widgetLabelTestUtils.ts, leaving only the describe/tests in the
spec (mirrors builderTestUtils pattern). Remove the dangling
research/architecture/toward-new-api-plan.md reference from renameWidget;
keep the ECS WidgetIdentity.label end-state intent. No fix-logic change.
